Add files via upload
In this Java program, the user inputs marks obtained (out of 100) in each subject. The program calculates the total marks by summing up the marks obtained in all subjects. Then, it calculates the average percentage by dividing the total marks by the total number of subjects. Based on the average percentage achieved, the program assigns grades. Finally, the program displays the total marks, average percentage, and the corresponding grade to the user.